hello_agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "fastapi>=0.115.12",
    "openai-agents>=0.0.9",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
